break state page into components
